Refactor models to ignore user when serializing
